fix: update branch of esp32-arduino-libs
The precompiled ESP-IDF libraries updated their branch on github.
The old one is not available anymore.
This leads to an error if you start from scratch or build with
a github action.
Locally you may need to remove the old one if you already checked it
out with platformio.

```
rm -r ~/.platformio/packages/framework-arduinoespressif32-libs
```