fix: set arduino lib version by git hash
force a specific version instead of only setting the branch.
When the branch name was used the requirement was fulfilled when any
version of this branch was installed.
This lead to a scenario where the continous integration used a newer,
incompatible version of this lib.

(cherry picked from commit bcac3ff72d990e3efbe62674f4ee4010f15de784)